Major refactor
Implement a popup stack. This stack allows for the enlarged popup images
to be seen above the preview popup - thus, when a popup is closed, only
the topmost popup is removed and the one that was below it is still
visible. This removes all special handling of what state the travel
popup is in, and makes keeping track of what popup (if any) are visible.
Clicking a navigation link removes all popups from the stack.